+++ services/app/lib/codebattle/game_process/engine/bot.ex
@@ -102,7 +102,8 @@ defmodule Codebattle.GameProcess.Engine.Bot do
 
     ActiveGames.terminate_game(game_id)
 
-    Notifications.notify_tournament("game:finished", fsm, %{
+    Notifications.notify_tournament(:game_over, fsm, %{
+      state: "finished",
       game_id: game_id,
       winner: {winner.id, "won"},
       loser: {loser.id, "lost"}
@@ -117,7 +118,8 @@ defmodule Codebattle.GameProcess.Engine.Bot do
     store_game_result!(fsm, {winner, "won"}, {loser, "gave_up"})
     ActiveGames.terminate_game(game_id)
 
-    Notifications.notify_tournament("game:finished", fsm, %{
+    Notifications.notify_tournament(:game_over, fsm, %{
+      state: "finished",
       game_id: game_id,
       winner: {winner.id, "won"},
       loser: {loser.id, "gave_up"}
