+++ services/app/lib/codebattle/game_process/engine/standard.ex
@@ -119,7 +119,8 @@ defmodule Codebattle.GameProcess.Engine.Standard do
     store_game_result!(fsm, {winner, "won"}, {loser, "lost"})
     ActiveGames.terminate_game(game_id)
 
-    Notifications.notify_tournament("game:finished", fsm, %{
+    Notifications.notify_tournament(:game_over, fsm, %{
+      state: "finished",
       game_id: game_id,
       winner: {winner.id, "won"},
       loser: {loser.id, "lost"}
@@ -133,7 +134,8 @@ defmodule Codebattle.GameProcess.Engine.Standard do
     store_game_result!(fsm, {winner, "won"}, {loser, "gave_up"})
     ActiveGames.terminate_game(game_id)
 
-    Notifications.notify_tournament("game:finished", fsm, %{
+    Notifications.notify_tournament(:game_over, fsm, %{
+      state: "finished",
       game_id: game_id,
       winner: {winner.id, "won"},
       loser: {loser.id, "gave_up"}
