+++ services/app/lib/codebattle/tournament/server.ex
@@ -1,9 +1,11 @@
 defmodule Codebattle.Tournament.Server do
-  def start(tournament) do
-    GenServer.start(__MODULE__, [tournament], name: tournament_key(tournament.id))
-  end
+  use GenServer
 
   # API
+  def start(tournament) do
+    GenServer.start(__MODULE__, tournament, name: tournament_key(tournament.id))
+  end
+
   def add_message(id, user, msg) do
     GenServer.cast(tournament_key(id), {:add_message, user, msg})
   end
@@ -17,38 +19,17 @@ defmodule Codebattle.Tournament.Server do
     end
   end
 
-  def update_tournament(id, event_type, params) do
-    GenServer.call(tournament_key(id), {:update_tournament, id, event_type, params})
+  def update_tournament(tournament_id, event_type, params) do
+    GenServer.cast(tournament_key(tournament_id), {event_type, params})
   end
 
   # SERVER
   def init(tournament) do
-    {:ok, %{tournament: tournament, messages: []}}
+    tournament_module = Codebattle.Tournament.Helpers.get_module(tournament)
+    {:ok, %{tournament: tournament, tournament_module: tournament_module, messages: []}}
   end
 
-  def handle_call({:update_tournament, id, "game:cancel", params}, _from, state) do
-    tournament =
-      id
-      |> Codebattle.Tournament.get!()
-      |> Codebattle.Tournament.Helpers.update_match(params.game_id, %{state: "canceled"})
-      |> Codebattle.Tournament.Helpers.maybe_start_new_step()
-
-    {:reply, tournament, state}
-  end
-
-  def handle_call({:update_tournament, id, "game:finished", params}, _from, state) do
-    tournament =
-      id
-      |> Codebattle.Tournament.get!()
-      |> Codebattle.Tournament.Helpers.update_match(
-        params.game_id,
-        Map.merge(params, %{state: "finished"})
-      )
-      |> Codebattle.Tournament.Helpers.maybe_start_new_step()
-
-    {:reply, tournament, state}
-  end
-
+  # Tournament chat
   def handle_call(:get_messages, _from, state) do
     %{messages: messages} = state
     {:reply, Enum.reverse(messages), state}
@@ -60,8 +41,27 @@ defmodule Codebattle.Tournament.Server do
     {:noreply, %{state | messages: new_msgs}}
   end
 
+  # Tournament
+  def handle_cast({event_type, params}, state) do
+    new_tournament = apply(state.tournament_module, event_type, [state.tournament, params])
+    broadcast_tournament(new_tournament)
+    {:noreply, Map.merge(state, %{tournament: new_tournament})}
+  end
+
+
   # HELPERS
+
+  defp broadcast_tournament(tournament) do
+    CodebattleWeb.Endpoint.broadcast!(
+      tournament_topic_name(tournament.id),
+      "update_tournament",
+      %{tournament: tournament}
+    )
+  end
+
   defp tournament_key(id) do
     {:via, :gproc, {:n, :l, {:tournament, "#{id}"}}}
   end
+
+  defp tournament_topic_name(tournament_id), do: "tournament_#{tournament_id}"
 end
