+++ services/app/lib/codebattle_web/controllers/tournament_controller.ex
@@ -17,13 +17,7 @@ defmodule CodebattleWeb.TournamentController do
   def show(conn, params) do
     tournament = Codebattle.Tournament.get!(params["id"])
 
-    view =
-      case tournament.type do
-        "team" -> CodebattleWeb.Live.Tournament.TeamView
-        _ -> CodebattleWeb.Live.Tournament.IndividualView
-      end
-
-    LiveView.Controller.live_render(conn, view,
+    LiveView.Controller.live_render(conn, CodebattleWeb.Live.Tournament.View,
       session: %{current_user: conn.assigns[:current_user], tournament: tournament}
     )
   end
