+++ services/app/lib/codebattle_web/live/tournament/view.ex
@@ -1,13 +1,13 @@
-defmodule CodebattleWeb.Live.Tournament.TeamView do
+defmodule CodebattleWeb.Live.Tournament.View do
   use Phoenix.LiveView
   use Timex
 
-  alias Codebattle.Tournament.Helpers
+  alias Codebattle.Tournament
 
   @update_frequency 1_000
 
   def render(assigns) do
-    CodebattleWeb.TournamentView.render("team.html", assigns)
+    CodebattleWeb.TournamentView.render("#{assigns.tournament.type}.html", assigns)
   end
 
   def mount(session, socket) do
@@ -16,7 +16,7 @@ defmodule CodebattleWeb.Live.Tournament.TeamView do
     end
 
     tournament = session.tournament
-    messages = Codebattle.Tournament.Server.get_messages(tournament.id)
+    messages = Tournament.Server.get_messages(tournament.id)
 
     CodebattleWeb.Endpoint.subscribe(topic_name(tournament))
 
@@ -57,93 +57,54 @@ defmodule CodebattleWeb.Live.Tournament.TeamView do
   end
 
   def handle_event("join", %{"team_id" => team_id}, socket) do
-    new_tournament =
-      Helpers.join(
-        socket.assigns.tournament,
-        socket.assigns.current_user,
-        String.to_integer(team_id)
-      )
+    Tournament.Server.update_tournament(socket.assigns.tournament.id, :join, %{
+      user: socket.assigns.current_user,
+      team_id: String.to_integer(team_id)
+    })
 
-    CodebattleWeb.Endpoint.broadcast_from(
-      self(),
-      topic_name(new_tournament),
-      "update_tournament",
-      %{
-        tournament: new_tournament
-      }
-    )
+    {:noreply, socket}
+  end
 
-    {:noreply, assign(socket, tournament: new_tournament)}
+  def handle_event("join", _params, socket) do
+    Tournament.Server.update_tournament(socket.assigns.tournament.id, :join, %{
+      user: socket.assigns.current_user
+    })
+
+    {:noreply, socket}
   end
 
   def handle_event("leave", _params, socket) do
-    new_tournament =
-      Helpers.leave(
-        socket.assigns.tournament,
-        socket.assigns.current_user
-      )
+    Tournament.Server.update_tournament(socket.assigns.tournament.id, :leave, %{
+      user: socket.assigns.current_user
+    })
 
-    CodebattleWeb.Endpoint.broadcast_from(
-      self(),
-      topic_name(new_tournament),
-      "update_tournament",
-      %{
-        tournament: new_tournament
-      }
-    )
-
-    {:noreply, assign(socket, tournament: new_tournament)}
+    {:noreply, socket}
   end
 
   def handle_event("cancel", _params, socket) do
-    new_tournament =
-      Helpers.cancel!(
-        socket.assigns.tournament,
-        socket.assigns.current_user
-      )
+    Tournament.Server.update_tournament(socket.assigns.tournament.id, :cancel, %{
+      user: socket.assigns.current_user
+    })
 
-    CodebattleWeb.Endpoint.broadcast_from(
-      self(),
-      topic_name(new_tournament),
-      "update_tournament",
-      %{
-        tournament: new_tournament
-      }
-    )
-
-    {:stop,
-     socket
-     |> put_flash(:info, "Tournament cancel successfully.")
-     |> redirect(to: "/tournaments")}
+    {:stop, redirect(socket, to: "/tournaments")}
   end
 
   def handle_event("start", _params, socket) do
-    new_tournament =
-      Helpers.start!(
-        socket.assigns.tournament,
-        socket.assigns.current_user
-      )
+    Tournament.Server.update_tournament(socket.assigns.tournament.id, :start, %{
+      user: socket.assigns.current_user
+    })
 
-    CodebattleWeb.Endpoint.broadcast_from(
-      self(),
-      topic_name(new_tournament),
-      "update_tournament",
-      %{
-        tournament: new_tournament
-      }
-    )
-
-    {:noreply, assign(socket, tournament: new_tournament)}
+    {:noreply, socket}
   end
 
-  # def handle_event("chat_message", %{"message" => %{"content" => ""}}, socket),
-  #   do: {:noreply, socket}
+  def handle_event("chat_message", %{"message" => %{"content" => ""}}, socket),
+    do: {:noreply, socket}
 
   def handle_event("chat_message", params, socket) do
     tournament = socket.assigns.tournament
     current_user = socket.assigns.current_user
 
-    Codebattle.Tournament.Server.add_message(
+    Tournament.Server.add_message(
       tournament.id,
       current_user,
       params["message"]["content"]
